temp

+++ gmailNotifyBot/Bot/DataBase/DataBaseModels/UserSettingsModel.cs
@@ -9,9 +9,9 @@ namespace CoffeeJelly.gmailNotifyBot.Bot.DataBase.DataBaseModels
 {
     public class UserSettingsModel
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
         public bool MailNotification { get; set; } = true;
